nodemap: remove v1 map
https://github.com/cilium/cilium/pull/31431 introduced v2 of the nodemap,
and as of v1.16 the datapath only uses this v2 map. Hence it's safe to
remove the agent logic that was still populating the v1 map.

Fixes: https://github.com/cilium/cilium/issues/34670
